+++ jupiter_swap/__init__.py
@@ -6,7 +6,7 @@ from .client import JupiterClient
 from .models import QuoteResponse, SwapResponse, TokenInfo, UltraOrder
 from .tokens import TokenClient
 
-__version__ = "0.1.1"
+__version__ = "0.2.0"
 
 __all__ = [
     "JupiterClient",